LayoutTests/inspector/console/console-dirxml.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function onload()
{
    var fragment = document.createDocumentFragment();
    fragment.appendChild(document.createElement("p"));

    console.dirxml(document);
    console.dirxml(fragment);
    console.dirxml(fragment.firstChild);
    console.log([fragment.firstChild]);

    runTest();
}

function test()
{
    InspectorTest.dumpConsoleMessages();
    InspectorTest.completeTest();
}
</script>
</head>

<body onload="onload()">
<p>
Tests that console logging dumps proper messages.
</p>

</body>
</html>
