LayoutTests/inspector/console/console-dirxml-expected.txt

CONSOLE MESSAGE: line 12: [object HTMLDocument]
CONSOLE MESSAGE: line 13: [object DocumentFragment]
CONSOLE MESSAGE: line 14: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 15: [object HTMLParagraphElement]
Tests that console logging dumps proper messages.

console-dirxml.html:12#document
console-dirxml.html:13Document Fragment
console-dirxml.html:14<p></p>
console-dirxml.html:15[<p></p>]
